# feat(react-icons,react-native-icons): improve build performance