--- a/.gitignore
+++ b/.gitignore
@@ -19,6 +19,8 @@ xcuserdata
 /ios/FluentIcons/Tests
 
 /packages/svg-icons/icons
+
+# react-icons
 packages/react-icons/intermediate
 packages/react-icons/react-icons-test/
 packages/react-icons/lib/
@@ -26,6 +28,12 @@ packages/react-icons/lib-cjs/
 packages/react-icons/src/icons
 packages/react-icons/src/sizedIcons
 packages/react-icons/src/index.tsx
+packages/react-icons/src/utils/fonts/*.ttf
+packages/react-icons/src/utils/fonts/*.woff
+packages/react-icons/src/utils/fonts/*.woff2
+packages/react-icons/src/utils/fonts/*.css
+packages/react-icons/src/utils/fonts/*.html
+packages/react-icons/src/utils/fonts/*.json
 packages/react-icons/yarn.lock
 packages/react-icons/package-lock.json