--- a/.gitignore
+++ b/.gitignore
@@ -28,6 +28,7 @@ packages/react-icons/lib-cjs/
 packages/react-icons/src/icons
 packages/react-icons/src/sizedIcons
 packages/react-icons/src/index.tsx
+packages/react-icons/src/fonts/
 packages/react-icons/src/utils/fonts/*.ttf
 packages/react-icons/src/utils/fonts/*.woff
 packages/react-icons/src/utils/fonts/*.woff2