--- a/packages/react-icons/src/utils/bundleIcon.styles.ts
+++ b/packages/react-icons/src/utils/bundleIcon.styles.ts
@@ -1,0 +1,7 @@
+import { makeStyles } from "@griffel/react";
+
+
+export const useBundledIconStyles = makeStyles({
+    root: { display: "none" },
+    visible: { display: "inline" }
+});