--- a/packages/react-icons/src/utils/useIconStyles.styles.ts
+++ b/packages/react-icons/src/utils/useIconStyles.styles.ts
@@ -1,0 +1,15 @@
+import { makeStyles } from "@griffel/react";
+
+export const useStyles = makeStyles({
+    root: {
+        display: 'inline',
+        lineHeight: 0,
+
+        "@media (forced-colors: active)": {
+          forcedColorAdjust: 'auto',
+        }
+    },
+    rtl : {
+      transform: 'scaleX(-1)'
+    }
+});