--- a/packages/react-native-icons/.gitignore
+++ b/packages/react-native-icons/.gitignore
@@ -1,1 +1,0 @@
-/intermediate