--- a/packages/react-native-icons/src/utils/fonts/.gitignore
+++ b/packages/react-native-icons/src/utils/fonts/.gitignore
@@ -1,1 +1,0 @@
-FluentSystemIcons*