… .github/workflows/bump-version.yml
… .github/workflows/cocoapods-publish.yml
… .github/workflows/pr.yml
… .github/workflows/publish.yml
… packages/react-icons/convert.js
… packages/react-icons/package.json
… packages/react-icons/src/.gitignore
… packages/react-icons/src/utils/useIconState.tsx
… packages/react-icons/tsconfig.generated-cjs.json
… packages/react-icons/tsconfig.json
… packages/react-icons/unfill-svg.js
… packages/react-native-icons/convert-font.js
… packages/react-native-icons/convert-rn.js
… packages/react-native-icons/package.json
… packages/react-native-icons/src/utils/fonts/createFluentFontIcon.tsx
… packages/react-native-icons/src/utils/wrapIcon.tsx
… packages/react-native-icons/tsconfig.json
… packages/react-native-icons/unfill-svg.js