needs: https://github.com/microsoft/fluentui-system-icons/pull/858

| task | Before | After |
|--------|--------|--------|
| `react-icons build` | 6m 43s | 1m 54s / 𝚫 72% FASTER |
| `react-native-icons build` | 5m 51s | 2m 12s / 𝚫 63% FASTER |

## general

- leverages task concurrency where possible
- makes build platform agnostic
- encapsulate resource heavy shell `unfill` task to nodejs script
- updates GHA(CI) to use secure version of actions + migrates to Node v20

## react-icons

- source files are transpiled via `tsc` for both esm and cjs
- generated content: 
  - whole chunks output is completely generated, including `.d.ts` files, to overcome huge perf implications
  - type definitions are encapsulated into single definition to make `tsc` processing faster 
  - tsc@5.6 is executed via `--no-check` mode to transpile generated esm (.js) files to cjs

## react-native-icons

- removes babel step completely as `tsc` handles jsx transpilation as expected